requirements.txt
pytest>=7.0.0
black>=23.0.0
fpdf2==2.7.7
arabic-reshaper==2.1.4
python-bidi==0.4.2
python-dotenv==1.0.1
google-generativeai==0.5.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
tqdm==4.66.2
